requirements.txt
functions-framework>=3.5.0,<4.0.0
flask>=3.0.2,<4.0.0
google-cloud-logging>=3.9.0,<4.0.0
langchain>=0.3.25,<0.4.0
langchain-openai>=0.3.17,<0.4.0
langgraph>=0.4.5,<0.5.0
python-dotenv>=1.1.0,<2.0.0
typing-extensions>=4.9.0,<5.0.0
asyncio>=3.4.3,<4.0.0 